Separate entity and range router
Take out to separate files logic for /elements and /elements/:id router
